# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Pay API",
        "description": "API de paiement mobile OM Pay pour la gestion bancaire",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentification Admin"
                ],
                "summary": "Connexion administrateur",
                "description": "Authentification des administrateurs pour accéder à la gestion des comptes",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email de l'administrateur",
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "a17b046a-0da9-3d17-ba26-978f3ea3a1e3"
                                                },
                                                "titulaire": {
                                                    "type": "string",
                                                    "example": "Administrateur Principal"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@gmail.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "admin"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants incorrects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Identifiants incorrects"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion avec code secret OM Pay",
                "description": "Authentification avec le code secret à 4 chiffres pour les connexions suivantes",
                "operationId": "aef2475c09e9a0443d079cd1d004b207",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "secret_code"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Numéro de téléphone",
                                        "type": "string",
                                        "example": "772687847"
                                    },
                                    "secret_code": {
                                        "description": "Code secret à 4 chiffres",
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "client": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "10281d44-9f2d-3008-8c61-7d55a71b538f"
                                                        },
                                                        "telephone": {
                                                            "type": "string",
                                                            "example": "772687847"
                                                        },
                                                        "user": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "string",
                                                                    "example": "a17b046a-0da9-3d17-ba26-978f3ea3a1e3"
                                                                },
                                                                "titulaire": {
                                                                    "type": "string",
                                                                    "example": "Abdoulaye Seck"
                                                                },
                                                                "email": {
                                                                    "type": "string",
                                                                    "example": "seckmoustapha238@gmail.com"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "compte_actif": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "62028976-381e-3a70-9073-df6de0fd5e74"
                                                        },
                                                        "num_compte": {
                                                            "type": "string",
                                                            "example": "C0081999796"
                                                        },
                                                        "solde": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 8728
                                                        },
                                                        "devise": {
                                                            "type": "string",
                                                            "example": "XOF"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "is_first_login": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Code secret incorrect ou compte bloqué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code secret incorrect"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtenu après authentification OM Pay. Format: Bearer {token}",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}